Refactor: "사용자 로그인 후 응답헤더 수정"
사용자 로그인 후 응답헤더 수정

Resolvers: #5
Ref: #3
