+++ src/main/java/com/devwinter/authservice/config/security/JwtTokenProvider.java
@@ -1,5 +1,6 @@
-package com.devwinter.authservice.config.security.handler;
+package com.devwinter.authservice.config.security;
 
+import com.devwinter.authservice.config.security.handler.TokenInfo;
 import com.devwinter.authservice.config.security.model.PrincipalUser;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
@@ -8,8 +9,9 @@ import io.jsonwebtoken.security.Keys;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
-import org.springframework.stereotype.Component;
 import org.springframework.security.core.GrantedAuthority;
+import org.springframework.stereotype.Component;
+
 import java.security.Key;
 import java.util.Date;
 import java.util.stream.Collectors;
@@ -17,10 +19,11 @@ import java.util.stream.Collectors;
 @Slf4j
 @Component
 public class JwtTokenProvider {
- 
+
     private final Key key;
     private final Long accessTokenValidTime;
     private final Long refreshTokenValidTime;
+
     public JwtTokenProvider(@Value("${jwt.secret-key}") String secretKey,
                             @Value("${jwt.access-token-expiration-time}") Long accessTokenValidTime,
                             @Value("${jwt.refresh-token-expiration-time}") Long refreshTokenValidTime) {
@@ -31,9 +34,10 @@ public class JwtTokenProvider {
     }
 
     public TokenInfo generateToken(Authentication authentication) {
-        String authorities = authentication.getAuthorities().stream()
-                .map(GrantedAuthority::getAuthority)
-                .collect(Collectors.joining(","));
+        String authorities = authentication.getAuthorities()
+                                           .stream()
+                                           .map(GrantedAuthority::getAuthority)
+                                           .collect(Collectors.joining(","));
 
         PrincipalUser principalUser = (PrincipalUser) authentication.getPrincipal();
 
@@ -43,23 +47,23 @@ public class JwtTokenProvider {
         Date accessTokenExpiresIn = new Date(now + accessTokenValidTime);
         String accessToken = Jwts.builder()
                                  .setSubject(principalUser.getEmail())
+                                 .setAudience(principalUser.getUserId().toString())
                                  .claim("auth", authorities)
                                  .setExpiration(accessTokenExpiresIn)
                                  .signWith(key, SignatureAlgorithm.HS256)
                                  .compact();
- 
+
         // Refresh Token 생성
         Date refreshTokenExpiresIn = new Date(now + refreshTokenValidTime);
         String refreshToken = Jwts.builder()
-                .setExpiration(refreshTokenExpiresIn)
-                .signWith(key, SignatureAlgorithm.HS256)
-                .compact();
- 
+                                  .setExpiration(refreshTokenExpiresIn)
+                                  .signWith(key, SignatureAlgorithm.HS256)
+                                  .compact();
+
         return TokenInfo.builder()
-                .grantType("Bearer")
-                .userId(principalUser.getUserId())
-                .accessToken(accessToken)
-                .refreshToken(refreshToken)
-                .build();
+                        .grantType("Bearer")
+                        .accessToken(accessToken)
+                        .refreshToken(refreshToken)
+                        .build();
     }
 }
