+++ src/main/java/com/devwinter/authservice/config/security/handler/TokenInfo.java
@@ -8,7 +8,6 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class TokenInfo {
-    private Long userId;
     private String grantType;
     private String accessToken;
     private String refreshToken;
